requirements.txt
pandas
numpy
scikit-learn
lightgbm
xgboost
ccxt
ta
pandas
numpy
scikit-learn
lightgbm
xgboost
ccxt
ta
joblib
tqdm
pyarrow

requests
python-dotenv
websocket-client
matplotlib
